pyproject.toml
[build-system]
requires = ["setuptools>65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fivegsim"
version = "0.1.0"
requires-python = ">=3.7"

dependencies = [
   "mocasin",
   "hydra-core",
   "numpy",
   "pandas",
]

[project.scripts]
fivegsim = "fivegsim.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["test", "*.test"]
namespaces = true

[tool.black]
line-length = 80
